2. Презентер для точек
